refactor(website): split PlanningOffice.svelte into sub-components
Extract PlanningOfficeDetail.svelte (shared form for desktop detail
and mobile sheet) and PlanningOfficeQueue.svelte (queue list with
drag-and-drop) to bring each file under the 500-line Svelte limit.

Reverts baseline.json PlanningOffice entry (no longer needed).
