+++ website/src/components/PlanningOffice.svelte
@@ -1,15 +1,9 @@
 <script lang="ts">
   import { onMount, onDestroy } from 'svelte';
+  import PlanningOfficeDetail from './PlanningOfficeDetail.svelte';
+  import PlanningOfficeQueue from './PlanningOfficeQueue.svelte';
 
   const { brand }: { brand: string } = $props();
-
-  const DOR_KEYS = ['spec_skizziert', 'offene_fragen_geklaert', 'abhaengigkeiten_klar', 'aufwand_geschaetzt'] as const;
-  const DOR_LABEL: Record<string, string> = {
-    spec_skizziert: 'Spec skizziert',
-    offene_fragen_geklaert: 'Fragen geklärt',
-    abhaengigkeiten_klar: 'Abhängigkeiten klar',
-    aufwand_geschaetzt: 'Aufwand geschätzt',
-  };
 
   interface PlanItem {
     extId: string;
@@ -195,15 +189,6 @@
     newDep = '';
   }
 
-  function priorityColor(p: string): string {
-    switch (p) {
-      case 'critical': return '#ef4444';
-      case 'high': return '#f97316';
-      case 'medium': return '#eab308';
-      default: return '#6b7280';
-    }
-  }
-
   onMount(() => {
     try {
       const v = localStorage.getItem('planungsbuero_view');
@@ -233,93 +218,35 @@
     <div class="pb-empty">Büro leer.</div>
   {:else}
     <div class="pb-layout" class:pb-mobile={isMobile}>
-      <div class="pb-queue" data-testid="pb-queue">
-        {#each items as it, idx (it.extId)}
-          <div
-            class="pb-row"
-            class:selected={!isMobile && selected?.extId === it.extId}
-            class:drag-source={dragSrcExtId === it.extId}
-            class:drop-target={dropTargetIdx === idx}
-            data-testid="pb-queue-row-{it.extId}"
-            draggable={!isMobile}
-            ondragstart={(e) => onDragStart(e, it)}
-            ondragover={(e) => onDragOver(e, idx)}
-            ondragleave={onDragLeave}
-            ondrop={(e) => onDrop(e, idx)}
-            ondragend={onDragEnd}
-            onclick={() => selectItem(it)}
-          >
-            <span
-              class="pb-handle"
-              onpointerdown={(e) => onHandlePointerDown(e, it)}
-              onpointermove={onHandlePointerMove}
-              onpointerup={onHandlePointerUp}
-            >☰</span>
-            <span class="pb-rank">{String(idx + 1).padStart(2, '0')}</span>
-            <span class="pb-priority-dot" style="background:{priorityColor(it.priority)}"></span>
-            <span class="pb-extid">{it.extId}</span>
-            <span class="pb-title">{it.title}</span>
-            <span class="pb-dor-squares">
-              {#each DOR_KEYS as k}
-                <span class="pb-dor-sq" class:pb-dor-on={it.readiness?.[k] === true}></span>
-              {/each}
-            </span>
-          </div>
-        {/each}
-      </div>
+      <PlanningOfficeQueue
+        {items}
+        {isMobile}
+        selectedExtId={selected?.extId ?? null}
+        {dragSrcExtId}
+        {dropTargetIdx}
+        onDragStart={onDragStart}
+        onDragOver={onDragOver}
+        onDragLeave={onDragLeave}
+        onDrop={onDrop}
+        onDragEnd={onDragEnd}
+        onSelect={selectItem}
+        onHandlePointerDown={onHandlePointerDown}
+        onHandlePointerMove={onHandlePointerMove}
+        onHandlePointerUp={onHandlePointerUp}
+      />
 
       {#if !isMobile && selected}
         <div class="pb-detail" data-testid="pb-detail">
-          <h2 class="pb-detail-title">{selected.title}</h2>
-          <label class="pb-field-label">Kern-Nutzen
-            <textarea
-              class="pb-textarea"
-              value={selected.valueProp ?? ''}
-              onblur={(e) => patch(selected!.extId, { valueProp: (e.target as HTMLTextAreaElement).value })}
-            ></textarea>
-          </label>
-          <fieldset class="pb-fieldset">
-            <legend>Definition of Ready</legend>
-            {#each DOR_KEYS as k}
-              <label class="pb-check">
-                <input type="checkbox" checked={selected.readiness?.[k] === true} onchange={() => toggleDor(selected!, k)} />
-                {DOR_LABEL[k]}
-              </label>
-            {/each}
-          </fieldset>
-          <div class="pb-deps">
-            <span class="pb-field-label">Abhängigkeiten</span>
-            <div class="pb-chips">
-              {#each selected.dependsOn as dep}
-                <span class="pb-chip">{dep}<button class="pb-chip-x" onclick={() => removeDep(dep)}>×</button></span>
-              {/each}
-            </div>
-            <input
-              class="pb-dep-input"
-              placeholder="Neue Abhängigkeit…"
-              bind:value={newDep}
-              onkeydown={(e) => { if (e.key === 'Enter') addDep(); }}
-            />
-          </div>
-          <div class="pb-effort-btns">
-            {#each ['klein', 'mittel', 'gross'] as eff}
-              <button
-                class="pb-effort-btn"
-                class:active={selected.effort === eff}
-                onclick={() => patch(selected!.extId, { effort: eff })}
-              >{eff}</button>
-            {/each}
-          </div>
-          <label class="pb-check pb-override-check">
-            <input type="checkbox" data-testid="pb-override" bind:checked={override} />
-            Override (trotz &lt; 4/4)
-          </label>
-          <button
-            class="pb-promote-btn"
-            data-testid="pb-detail-promote"
-            disabled={!override && selected.dorScore < 4}
-            onclick={() => promote(selected!)}
-          >Als nächstes planen</button>
+          <PlanningOfficeDetail
+            item={selected}
+            bind:override
+            bind:newDep
+            patchFn={patch}
+            toggleDorFn={toggleDor}
+            promoteFn={promote}
+            removeDepFn={removeDep}
+            addDepFn={addDep}
+          />
         </div>
       {/if}
     </div>
@@ -335,57 +262,16 @@
         <button class="pb-sheet-close" data-testid="pb-sheet-close" onclick={() => sheetOpen = false}>×</button>
       </div>
       <div class="pb-sheet-body">
-        <h2 class="pb-detail-title">{sheetItem.title}</h2>
-        <label class="pb-field-label">Kern-Nutzen
-          <textarea
-            class="pb-textarea"
-            value={sheetItem.valueProp ?? ''}
-            onblur={(e) => {
-              patch(sheetItem!.extId, { valueProp: (e.target as HTMLTextAreaElement).value });
-            }}
-          ></textarea>
-        </label>
-        <fieldset class="pb-fieldset">
-          <legend>Definition of Ready</legend>
-          {#each DOR_KEYS as k}
-            <label class="pb-check">
-              <input type="checkbox" checked={sheetItem.readiness?.[k] === true} onchange={() => toggleDor(sheetItem!, k)} />
-              {DOR_LABEL[k]}
-            </label>
-          {/each}
-        </fieldset>
-        <div class="pb-deps">
-          <span class="pb-field-label">Abhängigkeiten</span>
-          <div class="pb-chips">
-            {#each sheetItem.dependsOn as dep}
-              <span class="pb-chip">{dep}<button class="pb-chip-x" onclick={() => removeDep(dep)}>×</button></span>
-            {/each}
-          </div>
-          <input
-            class="pb-dep-input"
-            placeholder="Neue Abhängigkeit…"
-            bind:value={newDep}
-            onkeydown={(e) => { if (e.key === 'Enter') addDep(); }}
-          />
-        </div>
-        <div class="pb-effort-btns">
-          {#each ['klein', 'mittel', 'gross'] as eff}
-            <button
-              class="pb-effort-btn"
-              class:active={sheetItem.effort === eff}
-              onclick={() => patch(sheetItem!.extId, { effort: eff })}
-            >{eff}</button>
-          {/each}
-        </div>
-        <label class="pb-check pb-override-check">
-          <input type="checkbox" bind:checked={override} />
-          Override (trotz &lt; 4/4)
-        </label>
-        <button
-          class="pb-promote-btn"
-          disabled={!override && sheetItem.dorScore < 4}
-          onclick={() => promote(sheetItem!)}
-        >Als nächstes planen</button>
+        <PlanningOfficeDetail
+          item={sheetItem}
+          bind:override
+          bind:newDep
+          patchFn={patch}
+          toggleDorFn={toggleDor}
+          promoteFn={promote}
+          removeDepFn={removeDep}
+          addDepFn={addDep}
+        />
       </div>
     </div>
   {/if}
@@ -456,248 +342,10 @@
     grid-template-columns: 1fr;
   }
 
-  .pb-queue {
-    border-right: 1px solid var(--pb-border);
-    overflow-y: auto;
-  }
-
-  .pb-row {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-    min-height: 56px;
-    padding: 6px 12px;
-    cursor: pointer;
-    border-bottom: 1px solid var(--pb-border);
-    border-left: 3px solid transparent;
-    transition: background 0.12s;
-  }
-
-  .pb-row:hover {
-    background: var(--pb-surface-hover);
-  }
-
-  .pb-row.selected {
-    border-left: 3px solid var(--pb-amber);
-    background: var(--pb-selected-bg);
-  }
-
-  .pb-row.drag-source {
-    opacity: 0.4;
-  }
-
-  .pb-row.drop-target {
-    border-top: 2px solid var(--pb-amber);
-  }
-
-  .pb-handle {
-    cursor: grab;
-    color: var(--pb-text-muted);
-    font-size: 1rem;
-    width: 24px;
-    text-align: center;
-    user-select: none;
-    touch-action: none;
-  }
-
-  .pb-rank {
-    font-family: var(--pb-mono);
-    font-size: 0.75rem;
-    color: var(--pb-text-muted);
-    width: 20px;
-  }
-
-  .pb-priority-dot {
-    width: 8px;
-    height: 8px;
-    border-radius: 50%;
-    flex-shrink: 0;
-  }
-
-  .pb-extid {
-    font-family: var(--pb-mono);
-    font-size: 0.7rem;
-    color: var(--pb-amber);
-    white-space: nowrap;
-    min-width: 60px;
-  }
-
-  .pb-title {
-    flex: 1;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    font-size: 0.85rem;
-  }
-
-  .pb-dor-squares {
-    display: flex;
-    gap: 3px;
-    flex-shrink: 0;
-  }
-
-  .pb-dor-sq {
-    width: 6px;
-    height: 6px;
-    border-radius: 1px;
-    background: var(--pb-border);
-  }
-
-  .pb-dor-sq.pb-dor-on {
-    background: var(--pb-amber);
-  }
-
   .pb-detail {
     padding: 16px;
     overflow-y: auto;
     background: var(--pb-surface);
-  }
-
-  .pb-detail-title {
-    font-size: 1.1rem;
-    margin: 0 0 12px;
-    color: var(--pb-text);
-  }
-
-  .pb-field-label {
-    display: block;
-    font-size: 0.75rem;
-    color: var(--pb-text-muted);
-    margin-bottom: 4px;
-  }
-
-  .pb-textarea {
-    width: 100%;
-    min-height: 60px;
-    background: var(--pb-bg);
-    border: 1px solid var(--pb-border);
-    color: var(--pb-text);
-    border-radius: 4px;
-    padding: 6px 8px;
-    font-family: var(--pb-mono);
-    font-size: 0.8rem;
-    resize: vertical;
-    box-sizing: border-box;
-  }
-
-  .pb-fieldset {
-    border: 1px solid var(--pb-border);
-    border-radius: 4px;
-    padding: 8px 12px;
-    margin: 12px 0;
-  }
-
-  .pb-fieldset legend {
-    font-size: 0.75rem;
-    color: var(--pb-text-muted);
-  }
-
-  .pb-check {
-    display: flex;
-    align-items: center;
-    gap: 6px;
-    font-size: 0.8rem;
-    margin: 4px 0;
-    cursor: pointer;
-  }
-
-  .pb-check input[type="checkbox"] {
-    accent-color: var(--pb-amber);
-  }
-
-  .pb-deps {
-    margin: 12px 0;
-  }
-
-  .pb-chips {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 4px;
-    margin: 4px 0;
-  }
-
-  .pb-chip {
-    display: inline-flex;
-    align-items: center;
-    gap: 4px;
-    background: var(--pb-bg);
-    border: 1px solid var(--pb-border);
-    border-radius: 3px;
-    padding: 2px 6px;
-    font-size: 0.7rem;
-    color: var(--pb-amber);
-  }
-
-  .pb-chip-x {
-    background: none;
-    border: none;
-    color: var(--pb-text-muted);
-    cursor: pointer;
-    font-size: 0.8rem;
-    padding: 0;
-    line-height: 1;
-  }
-
-  .pb-chip-x:hover {
-    color: #ef4444;
-  }
-
-  .pb-dep-input {
-    width: 100%;
-    background: var(--pb-bg);
-    border: 1px solid var(--pb-border);
-    color: var(--pb-text);
-    border-radius: 4px;
-    padding: 4px 8px;
-    font-family: var(--pb-mono);
-    font-size: 0.75rem;
-    margin-top: 4px;
-    box-sizing: border-box;
-  }
-
-  .pb-effort-btns {
-    display: flex;
-    gap: 6px;
-    margin: 12px 0;
-  }
-
-  .pb-effort-btn {
-    background: var(--pb-bg);
-    border: 1px solid var(--pb-border);
-    color: var(--pb-text-muted);
-    padding: 4px 12px;
-    border-radius: 4px;
-    cursor: pointer;
-    font-family: var(--pb-mono);
-    font-size: 0.75rem;
-  }
-
-  .pb-effort-btn.active {
-    border-color: var(--pb-amber);
-    color: var(--pb-amber);
-    background: var(--pb-selected-bg);
-  }
-
-  .pb-override-check {
-    margin: 12px 0 8px;
-  }
-
-  .pb-promote-btn {
-    width: 100%;
-    padding: 8px;
-    background: var(--pb-amber);
-    color: var(--pb-bg);
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-family: var(--pb-mono);
-    font-size: 0.8rem;
-    font-weight: 600;
-  }
-
-  .pb-promote-btn:disabled {
-    opacity: 0.4;
-    cursor: not-allowed;
   }
 
   .pb-sheet {
@@ -740,10 +388,6 @@
   @media (max-width: 767px) {
     .pb-layout {
       grid-template-columns: 1fr;
-    }
-
-    .pb-queue {
-      border-right: none;
     }
   }
 </style>
